refactor: move ActiveTurnTracker from runtime/mod.rs to turn_scheduler.rs
